mcp_server/requirements.txt
mcp
fastmcp
psycopg2-binary
beautifulsoup4
python-dotenv

